python-backend/backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.1
numpy==1.26.4
Pillow==10.4.0
